Remove localization support
This resulted in unexpected errors, e.g. in the city database *not all*
entry fields had all the supported languages available. As it's a bit tricky
to choose an appropriate fallback behaviour on such cases, let's just
leave it to the caller to decide.

This means we now always return the database entries *as is*.