Add support for OTP 23
A few things required fixing.

The library wasn't compiling on OTP 23 due to the deprecation of
`http_uri:parse` together with having `warnings_as_errors` enabled.
Fixing this required a rather cumbersome conditional use of either
`http_uri:parse` or `uri_string:parse`, depending on OTP version.

HTTPS downloads broke for endpoints using a wildcard certificate.
This is because the validation of certificate names wasn't working
properly in OTP[1] up until recently.
Making it work with OTP 23 would have required changes to the
HTTPS certificate verification boilerplate, some of them dependent
on OTP version, and so I flipped the table and re-dit it all in a
separate library - `tls_certificate_check` - which `locus` now
imports and to which it delegates this important feature.

Finally, many test cases also broke because of changes to `httpd`'s
behaviour, which is used to serve database files to the different
test cases.
There was a nasty hardcoded relative path that should have been
redefined properly a long time ago. Well, it finally was.

[1]: see https://bugs.erlang.org/browse/ERL-1232