Restore the import of MaxMind test data by cloning its repo directly
There's some issue with the GH actions checkout plugin when it comes to
importing submodules.

It fails with the following error message:
```
Error: Input 'submodules' not supported when falling back to download
using the GitHub REST API. To create a local Git repository instead, add
Git 2.18 or higher to the PATH.
```

I'm not fighting this battle. Not today, anyway.

This reverts commit ec0eee2f90eacafbc1e3d9a7ee9443785521f7c9.