One more attempt at fixing the madness, the uncontrollable madness of importing test data
CI tests were running under an additional rebar3 profile, which meant
the test data -downloaded-as-a-raw-dependency ended up where we didn't
predict it would.

Since this distinction between CI tests and local tests is a remnant of
the time we put in place a bunch of workarounds for tests to pass on Travis,
let's drop almost all of it and see what happens.